+++ packages/@tinacms/styles/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/tinacms/tinacms/compare/@tinacms/styles@0.4.2-alpha.2...@tinacms/styles@0.4.2) (2020-04-27)
+
+**Note:** Version bump only for package @tinacms/styles
+
+
+
+
+
+## [0.4.2-alpha.2](https://github.com/tinacms/tinacms/compare/@tinacms/styles@0.4.2-alpha.1...@tinacms/styles@0.4.2-alpha.2) (2020-04-24)
+
+**Note:** Version bump only for package @tinacms/styles
+
+
+
+
+
 ## [0.4.2-alpha.1](https://github.com/tinacms/tinacms/compare/@tinacms/styles@0.4.2-alpha.0...@tinacms/styles@0.4.2-alpha.1) (2020-04-07)
 
 **Note:** Version bump only for package @tinacms/styles