+++ packages/react-tinacms-github/CHANGELOG.md
@@ -3,6 +3,119 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.3.0...react-tinacms-github@0.3.1) (2020-05-12)
+
+**Note:** Version bump only for package react-tinacms-github
+
+
+
+
+
+# [0.3.0](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.3.0-alpha.3...react-tinacms-github@0.3.0) (2020-05-11)
+
+
+### Bug Fixes
+
+* use-github-form accepts actions ([9fb83bf](https://github.com/tinacms/tinacms/commit/9fb83bf))
+
+
+
+
+
+# [0.3.0-alpha.3](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.3.0-alpha.2...react-tinacms-github@0.3.0-alpha.3) (2020-05-08)
+
+**Note:** Version bump only for package react-tinacms-github
+
+
+
+
+
+# [0.3.0-alpha.2](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.3.0-alpha.1...react-tinacms-github@0.3.0-alpha.2) (2020-05-08)
+
+**Note:** Version bump only for package react-tinacms-github
+
+
+
+
+
+# [0.3.0-alpha.1](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.3.0-alpha.0...react-tinacms-github@0.3.0-alpha.1) (2020-05-08)
+
+
+### Bug Fixes
+
+* callstack exceeded when accessing branchName ([6e2ebd3](https://github.com/tinacms/tinacms/commit/6e2ebd3)), closes [#1109](https://github.com/tinacms/tinacms/issues/1109)
+* export GithubError interface ([85fc71a](https://github.com/tinacms/tinacms/commit/85fc71a))
+* show error modal if creating a branch fails ([614bea6](https://github.com/tinacms/tinacms/commit/614bea6))
+
+
+
+
+
+# [0.3.0-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.2.0...react-tinacms-github@0.3.0-alpha.0) (2020-05-06)
+
+
+### Bug Fixes
+
+* **use-github-format-forms:** the options are optional ([88ccd92](https://github.com/tinacms/tinacms/commit/88ccd92))
+
+
+### Features
+
+* **github-client:** optionally accept scope ([880a6b6](https://github.com/tinacms/tinacms/commit/880a6b6))
+
+
+
+
+
+# [0.2.0](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.2.0-alpha.0...react-tinacms-github@0.2.0) (2020-05-04)
+
+
+### Bug Fixes
+
+* github-client branch name should default to base branch ([7ea6773](https://github.com/tinacms/tinacms/commit/7ea6773))
+
+
+
+
+
+# [0.2.0-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.1.0...react-tinacms-github@0.2.0-alpha.0) (2020-04-28)
+
+
+### Bug Fixes
+
+* persist return directory and filename ([ec837a2](https://github.com/tinacms/tinacms/commit/ec837a2))
+
+
+### Features
+
+* add lock icon ([f9df0f3](https://github.com/tinacms/tinacms/commit/f9df0f3))
+* added GithubMediaStore ([133cdcc](https://github.com/tinacms/tinacms/commit/133cdcc))
+* create new branches ([1c557a2](https://github.com/tinacms/tinacms/commit/1c557a2))
+* github.getBranchList ([2a91a61](https://github.com/tinacms/tinacms/commit/2a91a61))
+* support working from the origin repository ([18df649](https://github.com/tinacms/tinacms/commit/18df649))
+* support working from the origin repository ([9d7398e](https://github.com/tinacms/tinacms/commit/9d7398e))
+* switch between existing branches ([c1bf9a0](https://github.com/tinacms/tinacms/commit/c1bf9a0))
+
+
+
+
+
+# [0.1.0](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.1.0-canary3.2...react-tinacms-github@0.1.0) (2020-04-27)
+
+**Note:** Version bump only for package react-tinacms-github
+
+
+
+
+
+# [0.1.0-canary3.2](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.1.0-canary3.1...react-tinacms-github@0.1.0-canary3.2) (2020-04-24)
+
+**Note:** Version bump only for package react-tinacms-github
+
+
+
+
+
 # [0.1.0-canary3.1](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.1.0-canary3.0...react-tinacms-github@0.1.0-canary3.1) (2020-04-21)